+++ Ordenacao1/cpp/include/utils.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-int swap(std::vector<double> &vec, int i, int j);
+int swap(std::vector<int> &vec, int i, int j);
 
 
 #endif
