+++ Ordenacao1/cpp/src/merge_sort.cpp
@@ -4,7 +4,7 @@
 
 #include "utils.h"
 
-std::vector<double> merge_sort(std::vector<double> &vec)
+std::vector<int> merge_sort(std::vector<int> &vec)
 {
 
   if(vec.size() <= 1){
@@ -13,8 +13,8 @@ std::vector<double> merge_sort(std::vector<double> &vec)
 
   int half = vec.size() / 2;
 
-  std::vector<double> left(vec.begin(), vec.begin() + half);
-  std::vector<double> right(vec.begin() + half, vec.end());
+  std::vector<int> left(vec.begin(), vec.begin() + half);
+  std::vector<int> right(vec.begin() + half, vec.end());
 
   left = merge_sort(left);
   right = merge_sort(right);
@@ -23,9 +23,9 @@ std::vector<double> merge_sort(std::vector<double> &vec)
 
 }
 
-std::vector<double> merge(std::vector<double> &left, std::vector<double> &right)
+std::vector<int> merge(std::vector<int> &left, std::vector<int> &right)
 {
-  std::vector<double> vecres;
+  std::vector<int> vecres;
 
   auto ileft = left.begin();
   auto iright = right.begin();
