+++ Ordenacao1/cpp/src/quick_sort.cpp
@@ -4,7 +4,7 @@
 
 #include "utils.h"
 
-bool quick_sort(std::vector<double> &vec, int lo, int hi)
+bool quick_sort(std::vector<int> &vec, int lo, int hi)
 {
   if(lo < hi) {
     int p = quick_sort_partition(vec, lo, hi);
@@ -18,7 +18,7 @@ bool quick_sort(std::vector<double> &vec, int lo, int hi)
 
 }
 
-int quick_sort_partition(std::vector<double> &vec, int lo, int hi)
+int quick_sort_partition(std::vector<int> &vec, int lo, int hi)
 {
   int pivot = vec[hi];
   int i = lo;
